requirements.txt
# Core dependencies
sentence-transformers>=2.3.0
faiss-cpu==1.7.4
requests>=2.31.0
numpy==1.26.0
tqdm>=4.65.0
huggingface-hub>=0.16.0,<1.0

# Web framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data visualization
plotly>=5.17.0

# PDF export
fpdf2>=2.7.0

# Note: `cross-encoder` was removed due to PyPI availability issues.
# If you need CrossEncoder functionality, install it separately or
# provide a compatible wheel/repository. Example (experimental):
# python -m pip install git+https://github.com/UKPLab/cross-encoder.git
